requirements/development.txt
-r base.txt

django-test-plus==1.1.1
factory_boy==2.11.1
Faker==0.9.2
flake8==3.5.0
flake8-bugbear==18.2.0
flake8-docstrings==1.3.0
pep8-naming==0.7.0
pre-commit==1.10.4
psycopg2-binary==2.7.5
